:sparkles: lodash.debounce 제거, useDebounce 훅 구현하여 이벤트 최적화 적용
