+++ src/hooks/vote/index.ts
@@ -1,5 +1,7 @@
 export { useGetAllVotes } from './useGetAllVotes';
 export { useGetVoteById } from './useGetVoteById';
+export { useGetVoteBySearch } from './useGetVoteBySearch';
 export { default as useGetVoteReplies } from './useGetVoteReplies';
 export { default as useLikeVoteMutation } from './useLikeVoteMutation';
 export { default as useVotingMutation } from './useVotingMutation';
+
